requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
coverage==7.3.2
cryptography==41.0.5
defusedxml==0.8.0rc2
Django==4.2.7
django-debug-toolbar==4.2.0
django-debug-toolbar-force==0.2
django-filter==23.4
django-nine==0.2.7
django-rest-framework==0.1.0
djangorestframework==3.14.0
httpie==3.2.2
idna==3.6
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.4
oauthlib==3.2.2
packaging==23.2
psycopg2==2.9.9
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
PySocks==1.7.1
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.7.0
social-auth-app-django==5.4.0
social-auth-core==4.5.0
sqlparse==0.4.4
urllib3==2.1.0
